FEATURE - Finally add GameStateController spawner control based on heat level
-Sun reports to GameStateController when intensity changes
-GameStateController tries to activate spawners based on the new intensity level
-Defined new spawner prefabs (all with same unit type because why not) each with their own heat levels and unit sets
-Reduced spawn rates except for highest tier spawners

+++ Assets/Scripts/Units/Spawner/EnemySpawner.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class EnemySpawner : UnitSpawner
 {
     //threshold for heat level activation
-    public int HeatLevelThreshold;
+    public float HeatLevelThreshold;
 
     private MeshRenderer _editorViewMesh;
 
@@ -18,7 +18,7 @@ public class EnemySpawner : UnitSpawner
     }
 
     //activate the spawner if heat level is high enough
-    public void AttemptActivation(int heatLevel)
+    public void AttemptActivation(float heatLevel)
     {
         if(heatLevel >= HeatLevelThreshold)
         {
